fixed for watching parent folder + new events + tests

+++ src/tests/Jbta.SearchEngine.IntegrationTests/Utils/TestFile.cs
@@ -1,18 +1,28 @@
 ﻿using System;
 using System.IO;
+using System.Threading;
 
 namespace Jbta.SearchEngine.IntegrationTests.Utils
 {
     internal class TestFile : IDisposable
     {
-        public TestFile(string filePath, string content)
+        public TestFile(string content, string folderPath = ".", string fileName = null)
         {
-            Path = filePath;
-            using (var file = new StreamWriter(filePath, true))
+            fileName = fileName ?? GenerateFileName();
+            Path = $"{folderPath}\\{fileName}";
+            using (var file = new StreamWriter(Path, true))
             {
                 file.WriteLine(content);
             }
+            File.SetAttributes(Path, FileAttributes.Normal);
         }
+
+        public static string GenerateFileName()
+        {
+            return $"test-{Guid.NewGuid()}.txt";
+        }
+
+        public string Name => new FileInfo(Path).Name;
 
         public string Path { get; private set; }
 
@@ -31,9 +41,31 @@ namespace Jbta.SearchEngine.IntegrationTests.Utils
             Path = newPath;
         }
 
+        public void Move(string folderPath)
+        {
+            folderPath = folderPath.Replace(".\\", string.Empty);
+            var directoryPath = new FileInfo(Path).Directory.Name;
+            var newFilePath = Path.Replace(directoryPath, folderPath);
+            File.Move(Path, newFilePath);
+            Path = newFilePath;
+        }
+
         public void Dispose()
         {
+            Delete();
+        }
+
+        public void Delete()
+        {
+            if (!File.Exists(Path))
+            {
+                return;
+            }
+
             File.Delete(Path);
+            GC.Collect();
+            GC.WaitForPendingFinalizers();
+            Thread.Sleep(100);
         }
     }
 }
